Move updateTsconfig to a task file
